Give the compiler long double, which on System V means x87
The type was refused by name. It exists now on all three targets, and the
three do not agree about what it is: System V gives it x87 80-bit extended
format in sixteen bytes, while Apple arm64 and the UCRT both make it
another spelling of double. C90 asks only that it be no narrower than
double, so all three are conforming and only one needed a code generator.

x86_64-linux therefore gains a second floating unit. x87 is a stack rather
than a register file, so no long double is live in it between one
expression and the next: every value round-trips through memory the way an
SSE one does, sixteen bytes at a time. Arithmetic, all nine comparisons
with their NaN reasoning, both directions of conversion against every
other arithmetic type, ++ and --, constants, globals, parameters, returns
and va_arg are each written for it.

Two things it only learned from being run beside gcc:

- GNU as reverses fsub and fdiv against the Intel sense. Written the way
  the manual reads, 7.0L - 2.0L came out -5 and 7.0L / 2.0L came out
  0.2857. Add and multiply cannot ask the question, which is why half of
  it looked right.
- long double is class X87/X87UP, which is to say it is never in a
  register: it goes on the stack sixteen-aligned, and an aggregate holding
  one is MEMORY whatever its size - so struct { long double x; } is
  sixteen bytes, inside the limit that returns a struct in registers, and
  still comes back through the hidden pointer. glibc va_arg rounds the
  overflow pointer up to sixteen before reading one, so a long double left
  at an odd slot has printf read the neighbour.

arm64 needed one line and it was not the obvious one: double to long
double is a conversion the type system makes and the machine does not, and
asking the kinds rather than the registers emitted fcvt d0, d0, which the
assembler rejects. Only assembling finds that - -S alone does not.

Constants are taken apart with frexp rather than copied out of the host
long double, so a cc1 built where that type is 64 bits still writes a
correct 80-bit value; what such a host costs is the precision of the
constant, and that is now the one lossy thing about cross-compiling.

<float.h> gains the LDBL_ macros, which are the only per-target #ifdef in
that header.

tests/c90-probe.sh now reads 29 of 31. What is left is K&R definitions and
trigraphs, both declined rather than pending, and C23 deleted both.

+++ src/Ast.h
@@ -92,13 +92,16 @@ enum class BinOp { Add, Sub, Mul, Div, Mod, Shl, Shr,
 class Num final : public Expr {
 public:
     explicit Num(long v) : value_(v) {}
-    explicit Num(double d) : dvalue_(d) {}
+    // 'long double' rather than double, so that a constant folded at compile
+    // time keeps the width the target will give it. Narrowing happens once, in
+    // the backend, against the type the expression actually has.
+    explicit Num(long double d) : dvalue_(d) {}
     long value() const { return value_; }
-    double dvalue() const { return dvalue_; }
+    long double dvalue() const { return dvalue_; }
     void accept(Visitor &v) const override { v.visit(*this); }
 private:
     long value_ = 0;
-    double dvalue_ = 0;
+    long double dvalue_ = 0;
 };
 
 class Var final : public Expr {
